Reorganizing namespaces.
