+++ src/ForeverFactory/Builders/ManyBuilder.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using ForeverFactory.Builders.Common;
 using ForeverFactory.ExecutionContext;
 using ForeverFactory.Transforms;
 using ForeverFactory.Transforms.Conditions;
+using ForeverFactory.Transforms.Conditions.ExecutionContext;
 
 namespace ForeverFactory.Builders
 {
