--- src/ForeverFactory/ExecutionContext/IExecutionContext.cs
@@ -1,7 +1,0 @@
-﻿namespace ForeverFactory.ExecutionContext
-{
-    internal interface IExecutionContext // TODO rename
-    {
-        int QuantityToProduce { get; }
-    }
-}
