+++ src/ForeverFactory/Transforms/Conditions/ExecutionContext/IExecutionContext.cs
@@ -1,0 +1,7 @@
+﻿namespace ForeverFactory.Transforms.Conditions.ExecutionContext
+{
+    internal interface IExecutionContext
+    {
+        int QuantityToProduce { get; }
+    }
+}
