+++ src/ForeverFactory/Transforms/Conditions/ExecutionContext/InstanceSetExecutionContext.cs
@@ -1,6 +1,8 @@
-﻿namespace ForeverFactory.ExecutionContext
+﻿using ForeverFactory.Transforms.Conditions.ExecutionContext;
+
+namespace ForeverFactory.ExecutionContext
 {
-    internal class InstanceSetExecutionContext : IExecutionContext // TODO rename
+    internal class InstanceSetExecutionContext : IExecutionContext
     {
         public int QuantityToProduce { get; }
